requirements.txt
few_shot==0.3.0
ftfy==6.3.1
kornia==0.8.1
matplotlib==3.10.6
numpy==2.2.6
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==11.3.0
regex==2025.9.1
scikit_learn==1.7.2
scipy==1.15.3
setuptools==75.8.0
scikit-image
tabulate==0.9.0
thop==0.1.1.post2209072238
timm==1.0.20
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
tokenizers==0.21.0